Move all types to 1 file
